+++ config/customConfig.js
@@ -1,11 +1,23 @@
-'use strict';
+"use strict";
 
-const { customConfigPath, devCustomConfigPath, prodCustomConfigPath } = require('./paths');
-const fs = require('fs-extra');
+const {
+  customConfigPath,
+  devCustomConfigPath,
+  prodCustomConfigPath,
+  packageJsonPath
+} = require("./paths");
+const fs = require("fs-extra");
+const _ = require("lodash");
+const argv = require("yargs").argv;
 
 let customConfig = {};
 let devCustomConfig = {};
 let prodCustomConfig = {};
+let packageJsonConfig = {};
+
+if (fs.existsSync(packageJsonPath)) {
+  packageJsonConfig = require(packageJsonPath);
+}
 
 if (fs.existsSync(customConfigPath)) {
   customConfig = require(customConfigPath);
@@ -19,4 +31,21 @@ if (fs.existsSync(prodCustomConfigPath)) {
   prodCustomConfig = require(prodCustomConfigPath);
 }
 
-module.exports = { customConfig, devCustomConfig, prodCustomConfig };
+// process argv
+const customArgv = _.pick(argv, [
+  "cssModules",
+  "devHtmlTemplate",
+  "prodHtmlTemplate",
+  "bundleLibrary",
+  "library",
+  "libraryTarget"
+]);
+
+_.assign(customConfig, customArgv);
+
+module.exports = {
+  packageJsonConfig,
+  customConfig,
+  devCustomConfig,
+  prodCustomConfig
+};